requirements.txt
# requirements
pyarrow
grandiso
lark-parser
networkx
click
pypika
pyyaml
duckdb
toolz
jupyterlab
kuzu 
black 
